refactor: 에러 핸들링 및 요청 중단 커스텀 훅 구현
- fetchHandler 함수와 useAbortController 커스텀 훅을 통합
- 사용처에서 별도의 controller 생성 불필요
