+++ src/hooks/useFetch.ts
@@ -1,0 +1,34 @@
+import { AxiosError, AxiosResponse, isAxiosError } from 'axios';
+import { useAbortController } from './useAbortController';
+
+export function useFetch() {
+  const { createAbortController } = useAbortController();
+
+  const fetchHandler = async <T>(
+    fetchFn: (controller: AbortController) => Promise<AxiosResponse<T>>,
+    {
+      onSuccess,
+      onError,
+    }: {
+      onSuccess: (response?: AxiosResponse<T>) => void;
+      onError: (error?: AxiosError) => void;
+    },
+  ) => {
+    const controller = createAbortController();
+
+    try {
+      const response = await fetchFn(controller);
+      onSuccess(response);
+    }
+    catch (error) {
+      if (isAxiosError(error)) {
+        onError(error);
+      }
+      else {
+        console.error(error);
+      }
+    }
+  };
+
+  return { fetchHandler };
+}
